No PublicationTargetIdentifier for non-transactional listeners
For non-transactional listeners reliable event publication is not
necessary.
Relevant use-case for persistent events are domain events, and those
will always (?)
be handled by a transactional event listener marked with
@TransactionalEventListener.

Apart from that 'PublicationTargetIdentifier.forListener(listener)' will
throw an Exception
for all listeners that are not an instance of
'ApplicationListenerMethodAdapter'. This makes the project unusable in
spring boot projects with additional libraries such as spring-web or
spring-data-mongodb.